+++ phenodata/util.py
@@ -16,11 +16,23 @@ def setup_logging(level=logging.INFO):
         stream=sys.stderr,
         level=level)
 
-def normalize_options(options):
+def normalize_options(options, list_items=None):
     normalized = {}
+    list_items = list_items or []
     for key, value in options.items():
+
+        # Sanitize key
         key = key.strip('--<>')
+
+        # Decode list options
+        if key in list_items:
+            if value is None:
+                value = []
+            elif type(value) is str:
+                value = read_list(value)
+
         normalized[key] = value
+
     return normalized
 
 def to_list(obj):